# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.LinkSailorRequest:
    properties:
      sailor_id:
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        type: string
    type: object
...
  handler.RecordRaceRequest:
    properties:
      boat_class:
        example: Optimist
        type: string
      finish_time:
        example: 125.5
        type: number
      group_id:
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        type: string
      penalty:
        example: DSQ
        type: string
      sail_number:
        example: "10"
        type: string
      sailor_id:
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        type: string
    type: object
...
  handler.UpdateResultRequest:
    properties:
      finish_time:
        example: 125.5
        type: number
      penalty:
        example: DSQ
        type: string
      place:
        example: 1
        type: integer
    type: object
...
  model.Race:
    properties:
      boat_class:
        type: string
      created_at:
        type: string
      finish_time:
        type: number
      group_id:
        type: string
      id:
        type: string
      judge_id:
        type: string
      penalty:
        type: string
      place:
        type: integer
      sail_number:
        type: string
      sailor_id:
        type: string
      updated_at:
        type: string
    type: object
  model.RaceGroup:
    properties:
      category_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      number:
        type: integer
      regatta_id:
        type: string
      start_at:
        type: string
      status:
        $ref: '#/definitions/model.RaceGroupStatus'
      updated_at:
        type: string
    type: object
  model.RaceGroupStatus:
    enum:
    - scheduled
    - running
    - finished
    - cancelled
    type: string
    x-enum-varnames:
    - RGScheduled
    - RGRunning
    - RGFinished
    - RGCancelled
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Backend API for the sailing competition judging application. Handles
    regattas, race groups, individual race measurements, sailors, documents, and user
    authentication.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/user/judging-of-sailing
  title: Judging of Sailing API
  version: 1.0.0
paths:
...
  /race-groups:
    post:
      consumes:
      - application/json
      description: Creates a new race group (a single start session with multiple
        boats). Groups race measurements together.
      parameters:
      - description: Race group details
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.RaceGroup'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.RaceGroup'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create race group
      tags:
      - race-groups
  /race-groups/{group_id}/races:
    get:
      description: Returns all race measurements for a given race group, ordered by
        boat class and finish time.
      parameters:
      - description: Race group UUID
        in: path
        name: group_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Race'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List races by group
      tags:
      - races
  /race-groups/{id}:
    delete:
      description: Deletes a race group and all associated race measurements.
      parameters:
      - description: Race group UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete race group
      tags:
      - race-groups
    get:
      description: Returns a single race group by UUID.
      parameters:
      - description: Race group UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RaceGroup'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: race group not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get race group by ID
      tags:
      - race-groups
  /race-groups/{id}/finish:
    post:
      description: Changes the race group status to "finished", indicating the heat
        has ended and no more measurements can be added.
      parameters:
      - description: Race group UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Finish race group
      tags:
      - race-groups
  /race-groups/{id}/start:
    post:
      description: Changes the race group status to "running", indicating the heat
        has started.
      parameters:
      - description: Race group UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Start race group
      tags:
      - race-groups
  /races:
    post:
      consumes:
      - application/json
      description: Records a new individual race measurement for a boat. The judge_id
        is automatically set from the authenticated user. The sailor_id can be null
        if the sailor is not yet identified.
      parameters:
      - description: Race measurement details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RecordRaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Race'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Record race measurement
      tags:
      - races
  /races/{id}:
    delete:
      description: Deletes a race measurement by ID.
      parameters:
      - description: Race UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete race
      tags:
      - races
    get:
      description: Returns a single race measurement by UUID.
      parameters:
      - description: Race UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Race'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: race not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get race by ID
      tags:
      - races
  /races/{id}/link-sailor:
    post:
      consumes:
      - application/json
      description: Links a previously unlinked race measurement to a sailor. Used
        when the sailor was not known at measurement time.
      parameters:
      - description: Race UUID
        in: path
        name: id
        required: true
        type: string
      - description: Sailor ID to link
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LinkSailorRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Link race to sailor
      tags:
      - races
  /races/{id}/result:
    patch:
      consumes:
      - application/json
      description: Updates the finish time, place, and penalty for a race measurement.
      parameters:
      - description: Race UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated result fields
        in: body
        name: result
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateResultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update race result
      tags:
      - races
...
  /regattas/{regatta_id}/race-groups:
    get:
      description: Returns all race groups (heats) for a given regatta, ordered by
        group number.
      parameters:
      - description: Regatta UUID
        in: path
        name: regatta_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.RaceGroup'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List race groups by regatta
      tags:
      - race-groups
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token obtained from POST /api/v1/auth/login. Use format: Bearer
      <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"